remove compiler warning
git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@8943 060e22d8-9d1c-0410-868a-e3a950d63d95

+++ src/drt_f3/xfluid3_surface_evaluate.cpp
@@ -511,7 +511,7 @@ void DRT::ELEMENTS::XFluid3Surface::IntegrateSurfaceImpulsRate(
 //  const Teuchos::RCP<XFEM::InterfaceHandleXFSI> ih = params.get< Teuchos::RCP< XFEM::InterfaceHandleXFSI > >("interfacehandle",null);
     
   // check, if we have not enough dofs (intersected element or element in a hole)
-  if (lm.size() != iel*4)
+  if ((int)lm.size() != iel*4)
   {
     return;
   }
